Format the code
This change introduces formatting recommended in official style-guide. Couple
points from the documentation:

- Always have type annotations on top-level definitions.
- Always have 2 empty lines between top-level declarations.
- Always bring the body of the declaration down one line.

You can find more [here](http://elm-lang.org/docs/style-guide).

Something that is not mentioned here is difference of spacing applied for tuples
and for describing of application-precedence:

* There is additional space for tuples:

  ( arg1, arg2 )

* There is no additional space for binding function arguments

  myFun 1, (otherFun 2)

This makes reading code a little bit faster, consider this as an example:

  sendStartCall : Model -> Line.Callee -> ( Phoenix.Socket.Socket Msg, Cmd (Phoenix.Socket.Msg Msg) )

It's much faster to understand, that the function returns a tuple, where the
second element is a Cmd...