Read environment variables at runtime
Environment variables in config/* files will be normally be read and set at compile
time. When running in production with Docker, it is desireable to keep
these variables out of the image and pass them into the container at
runtime so that images can be shared and published without sharing secrets.